SGE job submission support #119
