+++ usamriidPathDiscov/helpers.py
@@ -162,12 +162,11 @@ def get_options():
                         metavar="JOBNAME",
                         type=str,
                         help="Target task(s) of pipeline.")
-    parser.add_argument("-j", "--jobs", dest="jobs",
-                        default=6,
-                        metavar="N",
+    parser.add_argument("--sge",
                         type=int,
-                        help="Allow N jobs (commands) to run "
-                             "simultaneously.")
+                        default=0,
+                        help="bool for sun grid engine (if 1,  use SGE qsub parallelization; if 0, no SGE; default is 0) "
+                             )
     parser.add_argument("-n", "--just_print", dest="just_print",
                         action="store_true", default=False,
                         help="Don't actually run any commands; just print "
@@ -196,6 +195,7 @@ def get_options():
     parser.add_argument('-R2', help ="Path to reverse fastq file")
     parser.add_argument('--param', action='store_true', help = "Generate sample param.txt file and edit after generating directory tree")
     parser.add_argument('--noparam', action='store_false', help = "Use the default param.txt file")
+    parser.add_argument('-c','--cpuNum', dest="cpuNum" ,default=10, type=int, help="Number of CPU to use, default is 10")
 
     # get help string
     f = StringIO.StringIO()
@@ -431,20 +431,6 @@ def setup_param(config):
         line = re.sub(r'TAX_NAMES', config['tax_names'], line.rstrip())
         print (line)
 
-def format_fastq(fastq_gz, fastq_output):
-    """Extract fastq.gz file
-
- Arguments:
-    - `fastq_file`: Gzipped fastq file downloaded from sequencing center
-    - `fastq_output`: gunzipped fastq file
-    """
-    import gzip
-    fo = open(fastq_output, 'w')
-    fh = gzip.open(fastq_gz, 'rb')
-    for line in fh:
-        fo.write(line)
-    fo.close()
-    fh.close()
 
 def isGzip(input):
     if input.endswith(".gz"):
