+++ usamriidPathDiscov/pathogen.pl
@@ -111,7 +111,6 @@ pod2usage(-verbose => 2) if ($man);
 # -------------------- main --------------------
 
 # set paths
-
 # scripts paths 
 $path_scripts=$RealBin;
