+++ usamriidPathDiscov/tasks.py
@@ -46,6 +46,40 @@ def copy_map_file(file_to_copy, file_copy):
 
     shutil.copyfile(file_to_copy, file_copy)
     return
+
+def format_fastq(fastq_gz, fastq_output):
+    """Extract fastq.gz file
+
+ Arguments:
+    - `fastq_file`: Gzipped fastq file downloaded from sequencing center
+    - `fastq_output`: gunzipped fastq file
+    """
+    import gzip
+    fo = open(fastq_output, 'w')
+    fh = gzip.open(fastq_gz, 'rb')
+    for line in fh:
+        fo.write(line)
+    fo.close()
+    fh.close()
+
+def sff_to_fastq(sff_file, fastq_output):
+    """Convert sff2fastq
+
+ Arguments:
+    - `sff_file`: sff file
+    - `fastq_output`:  fastq file
+    """
+    cmds = [
+        'sff2fastq',
+        sff_file,
+        fastq_output,
+    ]
+    cmds = '  '.join(cmds)
+    print cmds
+    p = runCommand(cmds, "F")
+    return
+
+
 def copyDir(file_to_copy, file_copy):
     """copy the mapfile to analyiss dir
 
@@ -101,7 +135,7 @@ def comment():
     return "*" * 80
 #pathogen.pl --sample sample1 --command step1 --paramfile param.txt --outputdir ../results/sample1 --R1 /my/data/R1.fastq.gz --R2 /my/data/R2.fastq.gz > ../logs/out1.o 2> ../logs/out1.e &
 #input=[pathDescTest/input/F.fastq, pathDescTest/input/R.fastq, pathDescTest, step1,pathDescTest/input/param.txt, pathDescTest/logs/out.o, pathDescTest/logs/out.e]
-def priStage(input, project_dir, paramFile,numreads, output):
+def priStage(input, project_dir, paramFile,numreads,sge, output):
     """run step  1 of pathogen discovery
     Arguments:
         -`input`: list of F and R fastq file
@@ -122,37 +156,73 @@ def priStage(input, project_dir, paramFile,numreads, output):
             ##'2>', output + "/step1/logs/out1.e",
 
             #]
-    if len(input)==2:
-        ffastq,rfastq = input
-        cmds = [
+    if sge == 1:
+
+        if len(input)==2:
+            ffastq,rfastq = input
+            cmds = [
+                'run_standard_stable4.pl',
+                '--sample', project_dir,
+                '--paramfile', paramFile,
+                '--outputdir', output,
+                '--R1', ffastq,
+                '--R2', rfastq,
+                '--SGE', str(sge),
+                '--blast_unassembled', str(numreads),
+            ]
+            cmds = '  '.join(cmds)
+            cmds += "| tee  -a "  + output + "/analysis.log"
+            print cmds
+            p = runCommand(cmds, "F")
+            return
+        else:
+            input =input
+            cmds = [
             'run_standard_stable4.pl',
             '--sample', project_dir,
             '--paramfile', paramFile,
             '--outputdir', output,
-            '--R1', ffastq,
-            '--R2', rfastq,
+            '--R1', input,
+            '--SGE', str(sge),
             '--blast_unassembled', str(numreads),
-        ]
-        cmds = '  '.join(cmds)
-        cmds += "| tee  -a "  + output + "/analysis.log"
-        print cmds
-        p = runCommand(cmds, "F")
-        return
+            ]
+            cmds = '  '.join(cmds)
+            cmds += "| tee  -a "  + output + "/analysis.log"
+            print cmds
+            p = runCommand(cmds, "F")
+            return
     else:
-        input =input
-        cmds = [
-        'run_standard_stable4.pl',
-        '--sample', project_dir,
-        '--paramfile', paramFile,
-        '--outputdir', output,
-        '--R1', input,
-        '--blast_unassembled', str(numreads),
-        ]
-        cmds = '  '.join(cmds)
-        cmds += "| tee  -a "  + output + "/analysis.log"
-        print cmds
-        p = runCommand(cmds, "F")
-        return
+           if len(input)==2:
+                ffastq,rfastq = input
+                cmds = [
+                    'run_standard_stable4.pl',
+                    '--sample', project_dir,
+                    '--paramfile', paramFile,
+                    '--outputdir', output,
+                    '--R1', ffastq,
+                    '--R2', rfastq,
+                    '--blast_unassembled', str(numreads),
+                ]
+                cmds = '  '.join(cmds)
+                cmds += "| tee  -a "  + output + "/analysis.log"
+                print cmds
+                p = runCommand(cmds, "F")
+                return
+           else:
+                input =input
+                cmds = [
+                'run_standard_stable4.pl',
+                '--sample', project_dir,
+                '--paramfile', paramFile,
+                '--outputdir', output,
+                '--R1', input,
+                '--blast_unassembled', str(numreads),
+                ]
+                cmds = '  '.join(cmds)
+                cmds += "| tee  -a "  + output + "/analysis.log"
+                print cmds
+                p = runCommand(cmds, "F")
+                return
